requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
streamlit>=1.39.0
